+++ src/com/intellij/idea/plugin/hybris/impex/highlighting/DefaultImpexSyntaxHighlighter.kt
@@ -76,15 +76,10 @@ class DefaultImpexSyntaxHighlighter : SyntaxHighlighterBase() {
             ImpexTypes.START_USERRIGHTS,
             ImpexTypes.END_USERRIGHTS -> USER_RIGHTS_KEYS
 
-            ImpexTypes.PASSWORD,
-            ImpexTypes.READ,
-            ImpexTypes.CHANGE,
-            ImpexTypes.CREATE,
-            ImpexTypes.DELETE,
-            ImpexTypes.REMOVE,
-            ImpexTypes.CHANGE_PERM, -> USER_RIGHTS_HEADER_PARAMETER_KEYS
+            ImpexTypes.PERMISSION, -> USER_RIGHTS_HEADER_PERMISSION_PARAMETER_KEYS
 
             ImpexTypes.TYPE,
+            ImpexTypes.PASSWORD,
             ImpexTypes.UID,
             ImpexTypes.MEMBEROFGROUPS,
             ImpexTypes.TARGET, -> USER_RIGHTS_HEADER_MANDATORY_PARAMETER_KEYS
@@ -138,7 +133,7 @@ class DefaultImpexSyntaxHighlighter : SyntaxHighlighterBase() {
         val ALTERNATIVE_PATTERN_KEYS: Array<TextAttributesKey> = pack(ImpexHighlighterColors.ALTERNATIVE_PATTERN)
         val DOCUMENT_ID_KEYS: Array<TextAttributesKey> = pack(ImpexHighlighterColors.DOCUMENT_ID)
         val USER_RIGHTS_KEYS: Array<TextAttributesKey> = pack(ImpexHighlighterColors.USER_RIGHTS)
-        val USER_RIGHTS_HEADER_PARAMETER_KEYS: Array<TextAttributesKey> = pack(ImpexHighlighterColors.USER_RIGHTS_HEADER_PARAMETER)
+        val USER_RIGHTS_HEADER_PERMISSION_PARAMETER_KEYS: Array<TextAttributesKey> = pack(ImpexHighlighterColors.USER_RIGHTS_HEADER_PARAMETER)
         val USER_RIGHTS_HEADER_MANDATORY_PARAMETER_KEYS: Array<TextAttributesKey> = pack(ImpexHighlighterColors.USER_RIGHTS_HEADER_MANDATORY_PARAMETER)
         val USER_RIGHTS_PERMISSION_ALLOWED_KEYS: Array<TextAttributesKey> = pack(ImpexHighlighterColors.USER_RIGHTS_PERMISSION_ALLOWED)
         val USER_RIGHTS_PERMISSION_DENIED_KEYS: Array<TextAttributesKey> = pack(ImpexHighlighterColors.USER_RIGHTS_PERMISSION_DENIED)